pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "harmonyemissions"
version = "0.3.0"
description = "Harmony of Emissions — configurable high-frequency laser-plasma emission sources (surface HHG, gas HHG, LWFA betatron)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Harmony of Emissions contributors" }]
keywords = [
    "laser-plasma",
    "high-harmonic-generation",
    "attosecond",
    "relativistic-oscillating-mirror",
    "betatron",
]
classifiers = [
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "xarray>=2023.1",
    "h5netcdf>=1.1",
    "h5py>=3.8",
    "matplotlib>=3.7",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "joblib>=1.3",
]

[project.optional-dependencies]
accel = ["numba>=0.58", "pyfftw>=0.13"]
gpu = ["cupy-cuda12x>=13.0"]
mpi = ["mpi4py>=3.1"]
all = ["harmonyemissions[accel,gpu,mpi]"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-benchmark>=4.0",
    "hypothesis>=6.0",
    "ruff>=0.1",
    "mypy>=1.5",
    "types-PyYAML>=6.0",
    "nbmake>=1.4",
    "jupyter>=1.0",
]

[project.scripts]
harmony = "harmonyemissions.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
harmonyemissions = [
    "data/*.yaml",
    "data/*.json",
    "detector/data/attenuation/*.json",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
# Mark-driven tiers. Default = fast suite.
addopts = "-m 'not slow and not gpu and not mpi and not benchmark'"
markers = [
    "slow: long-running tests (> ~1 s each)",
    "gpu: requires CUDA + cupy (set HARMONY_TEST_CUPY=1)",
    "mpi: requires mpi4py + mpirun (set HARMONY_TEST_MPI=1)",
    "benchmark: perf regression suite (pytest -m benchmark --benchmark-only)",
]

[tool.coverage.run]
source = ["src/harmonyemissions"]
branch = true
omit = [
    "*/backends/cupy.py",
    "*/backends/smilei.py",
    "*/backends/epoch.py",
    "*/parallel/mpi.py",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_configs = false
strict_optional = true
check_untyped_defs = true
exclude = ["tests/benchmarks/"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
extend-select = ["I", "UP", "B"]
ignore = ["B008", "UP015"]
